feat: Add the console.tron.log snippet
When we are in the development process, we use some consoles to debug,
and recently, thanks to Reactotron, we are using the console.tron.log.
So I use one snippet to add the console faster. Until this point, I use
'ctl' in my local machine, but I guess the 'tron' should be easier to
understand.