feat: Add some Saga Effects and Redux Snippets
When we use Redux in our app, we frequently use some code blocks, like
import the redux connect and dispatch, and link the component with
Redux. When we pass to use Sagas, more codes are added to list, calling
some effects with the yield method. So I created some personal snippets
to make those things fasters